+++ tests/full/test_full.tf
@@ -15,8 +15,8 @@ module "main" {
   source = "../.."
 
   node_id        = 111
-  pod            = 2
-  address        = "100.1.1.111/24"
+  pod_id         = 2
+  ip             = "100.1.1.111/24"
   gateway        = "100.1.1.254"
   endpoint_group = "OOB1"
 }