+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- BREAKING CHANGE: rename pod variable to pod_id
+- BREAKING CHANGE: rename address variable to ip
+
 ## 0.0.1
 
 - Initial release